backend/requirements.txt

flask==2.3.3
flask-cors==4.0.0


google-generativeai==0.3.2
firebase-admin==6.2.0
google-cloud-firestore==2.11.1


python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
Werkzeug==2.3.7
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
python-dateutil==2.8.2
jsonschema==4.19.2
typing-extensions==4.8.0
certifi==2023.7.22